Refactor AppBar and MobileMenu
